+++ internal/service/commit/count_commits.go
@@ -1,8 +1,11 @@
 package commit
 
 import (
-	"bufio"
+	"bytes"
 	"fmt"
+	"io/ioutil"
+	"strconv"
+	"time"
 
 	"gitlab.com/gitlab-org/gitaly/internal/helper"
 
@@ -24,26 +27,41 @@ func (s *server) CountCommits(ctx context.Context, in *pb.CountCommitsRequest) (
 		return nil, err
 	}
 
-	cmd, err := helper.GitCommandReader(ctx, "--git-dir", repoPath, "rev-list", string(in.GetRevision()))
+	cmdArgs := []string{"--git-dir", repoPath, "rev-list", "--count", string(in.GetRevision())}
+
+	if before := in.GetBefore(); before != nil {
+		cmdArgs = append(cmdArgs, "--before="+timestampToRFC3339(before.Seconds))
+	}
+	if after := in.GetAfter(); after != nil {
+		cmdArgs = append(cmdArgs, "--after="+timestampToRFC3339(after.Seconds))
+	}
+	if path := in.GetPath(); path != nil {
+		cmdArgs = append(cmdArgs, "--", string(path))
+	}
+
+	cmd, err := helper.GitCommandReader(ctx, cmdArgs...)
 	if err != nil {
 		return nil, grpc.Errorf(codes.Internal, "CountCommits: cmd: %v", err)
 	}
 	defer cmd.Kill()
 
-	count := 0
-	scanner := bufio.NewScanner(cmd)
-	for scanner.Scan() {
-		count++
-	}
-
-	if err := scanner.Err(); err != nil {
-		grpc_logrus.Extract(ctx).WithError(err).Info("ignoring scanner error")
-		count = 0
+	var count int64
+	countStr, readAllErr := ioutil.ReadAll(cmd)
+	if readAllErr != nil {
+		grpc_logrus.Extract(ctx).WithError(err).Info("ignoring git rev-list error")
 	}
 
 	if err := cmd.Wait(); err != nil {
 		grpc_logrus.Extract(ctx).WithError(err).Info("ignoring git rev-list error")
 		count = 0
+	} else if readAllErr == nil {
+		var err error
+		countStr = bytes.TrimSpace(countStr)
+		count, err = strconv.ParseInt(string(countStr), 10, 0)
+
+		if err != nil {
+			return nil, grpc.Errorf(codes.Internal, "CountCommits: parse count: %v", err)
+		}
 	}
 
 	return &pb.CountCommitsResponse{Count: int32(count)}, nil
@@ -55,4 +73,8 @@ func validateCountCommitsRequest(in *pb.CountCommitsRequest) error {
 	}
 
 	return nil
+}
+
+func timestampToRFC3339(ts int64) string {
+	return time.Unix(ts, 0).Format(time.RFC3339)
 }