+++ CHANGELOG.md
@@ -6,6 +6,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/210
 - Extend CountCommits RPC to support before/after/path arguments
   https://gitlab.com/gitlab-org/gitaly/merge_requests/252
+- Fix a bug in FindAllTags parsing lightweight tags
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/256
 
 v0.25.0